Refactor to create two different services to prefill applications from different sources
